+++ src/Builder.php
@@ -85,6 +85,10 @@ class Builder implements LoggerAwareInterface
         'Cecil\Step\Optimize\Js',
         'Cecil\Step\Optimize\Images',
     ];
+    /**
+     * Temporary directory name.
+     */
+    public const TMP_DIR = '.cecil';
 
     /**
      * Configuration object.
@@ -145,7 +149,6 @@ class Builder implements LoggerAwareInterface
      * Assets path collection.
      * This is an array that holds paths to assets (like CSS, JS, images) that are used in the build process.
      * It is used to keep track of assets that need to be processed or copied.
-     * It can be set to an array of paths or updated with new asset paths.
      * @var array
      */
     protected $assets = [];
@@ -185,17 +188,17 @@ class Builder implements LoggerAwareInterface
      */
     protected $generatorManager;
     /**
-     * Application version.
-     * @var string
-     */
-    protected static $version;
-    /**
      * Build metrics.
      * This array holds metrics about the build process, such as duration and memory usage for each step.
      * It is used to track the performance of the build and can be useful for debugging and optimization.
      * @var array
      */
     protected $metrics = [];
+    /**
+     * Application version.
+     * @var string
+     */
+    protected static $version;
     /**
      * Current build ID.
      * This is a unique identifier for the current build process.
@@ -204,7 +207,7 @@ class Builder implements LoggerAwareInterface
      * @var string
      * @see \Cecil\Builder::build()
      */
-    protected $buildId;
+    protected static $buildId;
 
     /**
      * @param Config|array|null    $config
@@ -259,7 +262,7 @@ class Builder implements LoggerAwareInterface
         $this->options = array_merge(self::OPTIONS, $options);
 
         // set build ID
-        $this->buildId = date('YmdHis');
+        self::$buildId = hash('adler32', date('YmdHis') . self::$version);
 
         // process each step
         $steps = [];
@@ -297,14 +300,6 @@ class Builder implements LoggerAwareInterface
         $this->getLogger()->notice(\sprintf('Built in %s (%s)', $this->metrics['total']['duration'], $this->metrics['total']['memory']));
 
         return $this;
-    }
-
-    /**
-     * Returns current build ID.
-     */
-    public function getBuildId(): string
-    {
-        return $this->buildId;
     }
 
     /**
@@ -480,17 +475,9 @@ class Builder implements LoggerAwareInterface
     }
 
     /**
-     * Set assets path list.
-     */
-    public function setAssets(array $assets): void
-    {
-        $this->assets = $assets;
-    }
-
-    /**
      * Add an asset path to assets list.
      */
-    public function addAsset(string $path): void
+    public function addToAssetsList(string $path): void
     {
         if (!\in_array($path, $this->assets, true)) {
             $this->assets[] = $path;
@@ -500,7 +487,7 @@ class Builder implements LoggerAwareInterface
     /**
      * Returns list of assets path.
      */
-    public function getAssets(): array
+    public function getAssetsList(): array
     {
         return $this->assets;
     }
@@ -582,6 +569,14 @@ class Builder implements LoggerAwareInterface
         }
 
         return self::$version;
+    }
+
+    /**
+     * Returns current build ID.
+     */
+    public static function getBuildId(): string
+    {
+        return self::$buildId;
     }
 
     /**