+++ src/Command/Build.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Cecil\Command;
 
+use Cecil\Builder;
 use Cecil\Util;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\Table;
@@ -104,7 +105,7 @@ EOF
         if ($input->getOption('output')) {
             $config['output']['dir'] = $input->getOption('output');
             if ($input->getOption('output') != self::SERVE_OUTPUT) {
-                Util\File::getFS()->dumpFile(Util::joinFile($this->getPath(), self::TMP_DIR, 'output'), (string) $input->getOption('output'));
+                Util\File::getFS()->dumpFile(Util::joinFile($this->getPath(), Builder::TMP_DIR, 'output'), (string) $input->getOption('output'));
             }
         }
         if ($input->getOption('optimize') === true) {